pyproject.toml
[project]
name = "cover-letter-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=0.30.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "requests>=2.32.4",
    "streamlit>=1.47.1",
]
